requirements.txt
pan-os-python==1.2.0
pandevice==0.14.0
PyYAML==5.4.1
